+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 ### BREAKING CHANGES
 -->
 
+<a name="1.1.0"></a>
+
+## 1.1.0 (2017-05-09)
+
+### Features
+
+* Added abstract optimistic concurrency control entities (`*OccEntity.java`)
+* There are DTOs for all entities
+
+### BREAKING CHANGES
+
+* Renamed all entities
+
 <a name="1.0.0"></a>
 
 ## 1.0.0 (2017-04-28)